Split TcpEventBusBridgeHandler out of TcpEventBusBridgeImpl
